+++ appcli/commands/backup_manager_cli.py
@@ -74,7 +74,16 @@ class BackupManagerCli:
             cli_context.get_configuration_dir_state().verify_command_allowed(
                 AppcliCommand.BACKUP
             )
-            services_cli = cli_context.commands["service"]
+
+            try:
+                services_cli = cli_context.commands["service"]
+            except KeyError:
+                # The `service` command is not available, which means the orchestrator has no services to manage.
+                # Therefore there are no services to stop.
+                logger.info("No services to stop.")
+                services_cli = None
+                pre_stop_services = False
+                post_start_services = False
 
             if pre_stop_services:
                 logger.info("Stopping application services ...")
@@ -135,7 +144,15 @@ class BackupManagerCli:
             if not backup_manager.backup_file_exists(ctx, backup_file):
                 error_and_exit(f"Backup file [{backup_file}] not found.")
 
-            services_cli = cli_context.commands["service"]
+            try:
+                services_cli = cli_context.commands["service"]
+            except KeyError:
+                # The `service` command is not available, which means the orchestrator has no services to manage.
+                # Therefore there are no services to stop.
+                logger.info("No services to stop.")
+                services_cli = None
+                pre_stop_services = False
+                post_start_services = False
 
             if pre_stop_services:
                 logger.info("Stopping application services ...")